pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["Kkkasum <kasum2605@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.4.1"
pydantic-settings = "^2.2.1"
pymongo = "^4.6.2"
python-dateutil = "^2.9.0.post0"
motor = "^3.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
